requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
pip==21.3.1
setuptools==59.2.0
sqlparse==0.4.2
tzdata==2021.5
wheel==0.37.0
django-environ==0.8.1
whitenoise==5.3.0
pymssql==2.2.2